requirements.txt
transformers[torch]
torch
streamlit
python-dotenv
sentencepiece
numpy
pandas